remove 1.1 docs from the site
We never ran black for 1.1 nor did we add pre-commit hooks,
which now makes it very hard to work with that branch because
the pre-commit hooks in my working directory won't even let me commit
without a special environment variable.   as I am hoping to merge
a huge series of link corrections to the documentation for all
published branches, it would be very difficult to get these changes
to backport to 1.1 without it being brought up to date.
1.1 is EOL so it should be OK to remove from the site.