make the links to "Documentation" exceedingly obvious
People are apparently put off by terms like "Library"
instead of "Documentation" as well as the wordiness
of the front of the site, as the profile of SQLAlchemy's "new user"
has shifted in the past 16 years.

Originally, they were "evaluators";
Python developers, typically in architectural roles, looking
for new SQL tools to use for their projects.  The site had
to make a pitch and a case for SQLAlchemy and its philosophies
and rationale, hoping to attract these people as new users.

Today, with the site's management of those "evaluators" being
overwhelmingly successful, new users are no longer
"evaluators".  They arrive already wanting to get started.
The typical new user is now someone who is already tasked with using
SQLAlchemy, either on a job, in a college course, or if
they are in an "evaluation" role they are coming to us
based on our existing reputation.  They are also much more likely
to be not entirely happy about having to deal with SQLAlchemy
at all; a strange turn for an open source project to be so
popular and successful that it now has users who hate it but
are forced to use it, something I certainly never imagined
or wanted, any more than any creator would want anyone
to be "forced" to consume their art.  But of course
that's what happens because this is a tool used by employers
and institutions.  The entire reason SQLAlchemy has "Core"
separate from "ORM" is that I never wanted anyone to
feel like they "had" to use the ORM.  But that's happened
anyway.

Additionally, the bar for open source websites has changed,
where textual descriptions
and rationales etc. are not that interesting anymore and the
most successful projects have few textual elements as users
are guided to getting started; that is, users coming
to the site don't need to be "sold" anymore they just want
to "get started".  The sales pitch for something new is
just whatever pizazz / buzz they can create enough to get
someone to do a walkthrough video and that's how it goes.

SQLAlchemy's old, and its creator can talk talk talk
forever, so we're not going to win that game but at least
let's make "DOCUMENTATION" exceedingly obvious and remove
any arty terms like "Library" that seem to have been throwing
people off.   Obviously the docs themselves are still huge
but making them more digestible is an ongoing work in progress.

Any HTML errors I might have made here we can fix in a subsequent
bump.